+++ backend/footprint/src/main/java/foot/footprint/domain/comment/application/edit/EditCommentService.java
@@ -1,0 +1,6 @@
+package foot.footprint.domain.comment.application.edit;
+
+public interface EditCommentService {
+
+    public void edit(Long commentId, Long memberId, String newContent);
+}
